# Conflicts:
#	Backend/user-center-backend/src/main/java/com/leese/usercenter/controller/order/OrderController.java
#	Backend/user-center-backend/src/main/resources/generator/mapper/OrderMapper.xml
#	Fronted/user-center-fronted/src/api/order.ts
#	Fronted/user-center-fronted/src/page/order/OrderCustomerView.vue
#	Fronted/user-center-fronted/src/page/order/OrderList.vue

+++ Backend/user-center-backend/src/main/java/com/leese/usercenter/mod5/model/dto/ReviewResponse.java
@@ -1,21 +1,17 @@
 package com.leese.usercenter.mod5.model.dto;
 
 import lombok.Data;
-
 import java.util.Date;
 import java.util.List;
 
-/**
- * 评价响应
- */
 @Data
 public class ReviewResponse {
     private Long reviewId;
-    private Long orderId;
-    private Long userId;
+    private Integer orderId;
+    private Integer userId;
     private String userName;
-    private Integer rating;  // 改为 rating
-    private String comment;  // 改为 comment
+    private Integer rating;
+    private String comment;
     private Date reviewTime;
     private List<DishRatingResponse> dishRatings;
