+++ Backend/user-center-backend/src/main/java/com/leese/usercenter/mod5/model/entity/Review.java
@@ -1,74 +1,37 @@
 package com.leese.usercenter.mod5.model.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
-
 import java.io.Serializable;
 import java.util.Date;
 
-/**
- * 订单评价（对应现有review表）
- */
 @Data
 @TableName(value = "review")
 public class Review implements Serializable {
-    /**
-     * id
-     */
     @TableId(type = IdType.AUTO)
     private Long id;
 
-    /**
-     * 用户id
-     */
     @TableField(value = "userId")
-    private Long userId;
+    private Integer userId;
 
-    /**
-     * 订单id
-     */
     @TableField(value = "orderId")
-    private Long orderId;
+    private Integer orderId;
 
-    /**
-     * 评分（1-5星）
-     */
     private Integer rating;
 
-    /**
-     * 评价内容
-     */
     private String comment;
 
-    /**
-     * 点赞数
-     */
     private Integer likes;
 
-    /**
-     * 点赞用户（这个字段可能不需要，简化处理）
-     */
     @TableField(value = "likesUserId")
-    private Long likesUserId;
+    private Integer likesUserId;
 
-    /**
-     * 创建时间
-     */
     @TableField(value = "createTime")
     private Date createTime;
 
-    /**
-     * 更新时间
-     */
     @TableField(value = "updateTime")
     private Date updateTime;
 
-    /**
-     * 是否删除
-     */
     @TableField(value = "isDelete")
     private Integer isDelete;
